rearranging splitting factors

+++ src/apfelcomb/fk_qcd.h
@@ -115,7 +115,7 @@ namespace QCD
   // APFEL FK functions
   void avals(const int& xi, const double& xo, const int& fi, const double& Q, double*);
   void avals_pbar(const int& xi, const double& xo, const int& fi, const double& Q, double*);
-  double bvals(int const& xo, int const& xi, int const& fo, int const& fi);
+  double bvals(double const& xo, int const& xi, int const& fo, int const& fi);
 
   double diskernel(std::string const& obs, double const& x, double const& Q, double const& y, int const& i, int const& beta);
   double disobs(std::string const& obs, double const& x, double const& Q, double const& y);
